requirements.txt
numpy==2.2.1
pandas==2.2.3
matplotlib==3.10.0
alpaca_trade_api==3.2.0
alpaca-py==0.35.0
datetime==5.5
colorama==0.4.6
python-dotenv==1.0.1
flask==3.1.0
flask_cors==5.0.0
flask-socketio==5.5.1